More command tightening
