+++ source/commands/danger-process.ts
@@ -1,13 +1,11 @@
 import * as chalk from "chalk"
 import * as program from "commander"
 
-import { getCISource } from "../ci_source/get_ci_source"
-import { providers } from "../ci_source/providers"
 import { getPlatformForEnv } from "../platforms/platform"
-import { sentence } from "../runner/DangerUtils"
 import { Executor } from "../runner/Executor"
 import runDangerSubprocess, { prepareDangerDSL } from "./utils/runDangerSubprocess"
 import setSharedArgs, { SharedCLI } from "./utils/sharedDangerfileArgs"
+import getRuntimeCISource from "./utils/getRuntimeCISource"
 
 // Given the nature of this command, it can be tricky to test, so I use a command like this:
 //
@@ -34,18 +32,13 @@ program.action(process_name => (subprocessName = process_name)).parse(process.ar
 // use this to work with dynamic propeties
 const app = (program as any) as SharedCLI
 
-process.on("unhandledRejection", function(reason: string, _p: any) {
-  console.log(chalk.red("Error: "), reason)
-  process.exitCode = 1
-})
-
 if (process.env["DANGER_VERBOSE"] || app.verbose) {
   global.verbose = true
 }
 
 // a dirty wrapper to allow async functionality in the setup
-async function run(): Promise<any> {
-  const source = await setupCISource()
+async function run() {
+  const source = await getRuntimeCISource(app)
 
   // This does not set a failing exit code
   if (source && !source.isPR) {
@@ -81,26 +74,6 @@ async function run(): Promise<any> {
       }
     }
   }
-}
-
-const setupCISource = async () => {
-  const source = getCISource(process.env, app.externalCiProvider || undefined)
-
-  if (!source) {
-    console.log("Could not find a CI source for this run. Does Danger support this CI service?")
-    console.log(`Danger supports: ${sentence(providers.map(p => p.name))}.`)
-
-    if (!process.env["CI"]) {
-      console.log("You may want to consider using `danger pr` to run Danger locally.")
-    }
-
-    process.exitCode = 1
-  }
-  // run the sources setup function, if it exists
-  if (source && source.setup) {
-    await source.setup()
-  }
-  return source
 }
 
 run()
