--- source/commands/danger-runner.ts
@@ -1,1 +1,0 @@
-// Same as the other one but instead `import` the JS into the current runtime, so that it runs safely
